-ref: moved the TableSearch component from the Table component to its own file -ref: replaced var for const and let
